top spenders portion complete!
gosh, this was more complicated. I definitely had to rethink how to make a dataframe with this, and ended up using the tuple way because for some reason the other ways were not working? I am not sure why, but I did it in the end! I also refined some other portions of my previous codes to be more concise/efficient